@@ -3,58 +3,40 @@
   "version": "0.0.0",
   "description": "Jan Kuri's personal homepage",
   "license": "MIT",
+  "type": "module",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --open",
-    "build-dev": "ng build --configuration production",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "dev:ssr": "ng run jankuri-me:serve-ssr",
-    "serve:ssr": "node dist/server/main.js",
-    "build:ssr": "ng build --configuration production && ng run jankuri-me:server:production",
-    "prerender": "ng run jankuri-me:prerender",
-    "vercel-build": "yarn run build:ssr"
+    "build": "ng build",
+    "vercel-build": "npm run build"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^16.2.10",
-    "@angular/common": "^16.2.10",
-    "@angular/compiler": "^16.2.10",
-    "@angular/core": "^16.2.10",
-    "@angular/forms": "^16.2.10",
-    "@angular/platform-browser": "^16.2.10",
-    "@angular/platform-browser-dynamic": "^16.2.10",
-    "@angular/platform-server": "^16.2.10",
-    "@angular/router": "^16.2.10",
+    "@angular/animations": "^17.0.2",
+    "@angular/common": "^17.0.2",
+    "@angular/compiler": "^17.0.2",
+    "@angular/core": "^17.0.2",
+    "@angular/forms": "^17.0.2",
+    "@angular/platform-browser": "^17.0.2",
+    "@angular/platform-browser-dynamic": "^17.0.2",
+    "@angular/platform-server": "^17.0.2",
+    "@angular/router": "^17.0.2",
+    "@angular/ssr": "^17.0.0",
     "@fortawesome/fontawesome-free": "^6.0.0",
-    "@nguniversal/express-engine": "^16.2.0",
     "bulma": "^0.9.3",
     "d3": "^7.3.0",
     "express": "^4.17.3",
     "rxjs": "~6.6.0",
     "tslib": "^2.3.1",
-    "zone.js": "~0.13.3"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^16.0.1",
-    "@angular-devkit/build-angular": "^16.2.7",
-    "@angular/cli": "^16.2.7",
-    "@angular/compiler-cli": "^16.2.10",
-    "@nguniversal/builders": "^16.2.0",
+    "@angular-devkit/build-angular": "^17.0.0",
+    "@angular/cli": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.2",
     "@types/d3": "^7.1.0",
     "@types/express": "^4.17.13",
-    "@types/jasmine": "~3.10.3",
     "@types/node": "^20.8.7",
-    "jasmine-core": "~4.0.0",
-    "jasmine-marbles": "~0.8.4",
-    "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "purgecss-webpack-plugin": "^4.1.3",
-    "typescript": "~4.9.5"
+    "typescript": "~5.2.2"
   }
 }
