@@ -1,10 +1,7 @@
-import { enableProdMode } from '@angular/core';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { AppComponent } from './app/app.component';
+import { config } from './app/app.config.server';
 
-import { environment } from './environments/environment';
+const bootstrap = () => bootstrapApplication(AppComponent, config);
 
-if (environment.production) {
-  enableProdMode();
-}
-
-export { AppServerModule } from './app/app.server.module';
-export { renderModule } from '@angular/platform-server';
+export default bootstrap;
